@@ -2,7 +2,7 @@
 using System;
 using System.ComponentModel;
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 
 namespace Dapper.Tests.Performance
 {
