@@ -1249,7 +1249,6 @@ namespace Dapper
             return Parse<T>(result);
         }
 
-#if NET5_0_OR_GREATER
         /// <summary>
         /// Execute a query asynchronously using <see cref="IAsyncEnumerable{dynamic}"/>.
         /// </summary>
@@ -1338,12 +1337,15 @@ namespace Dapper
                             try { cmd?.Cancel(); }
                             catch { /* don't spoil any existing exception */ }
                         }
+#if NET5_0_OR_GREATER
                         await reader.DisposeAsync();
+#else
+                        reader.Dispose();
+#endif
                     }
                     if (wasClosed) cnn.Close();
                 }
             }
         }
-#endif
     }
 }
